Change from relative to absolute time in the output grib files
ungrib.exe is not able to read correctly realtive timestamp. Setting time units to "hours" (cdo -settunits,hours) fixes the problem.
Variable noleap is read from a BCtable. This variabels controls if the leap year will be removed or not:
For the evaluation runs typically is is not removed (false), and for the historical and scenario runs it is (true) in case the GCM uses "standard" calendar.