Removed "set -u" check
This option affects variables. If variable not defined/used, error is produced, which causes the program to immediately exit.
For example, if a variable that represents land_mask (sftlf) is not processed, the script will exit producing an error "land_mask: unbound variable".
To avoid this, "set -u" has been removed.