Changed psl interpolation filter
Interpolation filter for psl changed from fixed2interval to tinterp2interval. fixed2interval keeps only the first and the last timestep,
and interpolates the steps between. Therefore it is appropriate only for the fixed variables and those that are used only for the initilazation such as soil moisture and soil temperature.
tinterp2interval keeps all the exixting timesteps, and interpolates those that are missing.