Functionality upgrade due to cdo and nco version update, added scripts for direct GCM data download (#4)
This is a squash commit from to the devel branch on Yoselita/bctool, including:

* Temporal interpolation of fixed files included. For SST tos instead of tslsi.

* nco and cdo commands adjusted to work correctly with the newer versions of cdo (tested with 1.9.8) and nco (tested with 4.9.4).

* Height of the surface variables set to 0, to be included at the first vertical level.

* An option to choose interpolation interval in time included in the header (interval). Names for temporal interpolation filters changed.

* Ungrib runs only 1 time as fixed and soil variabels are to be interpolated at the requested frequency. Reading and the function for updating namelists changed - includes option to read different intial and ending times.

* SM and ST at the deepest layer for the CanESM2 model included (40-405m). Interpolation techinques added for SST inteprolation.

* Order of the functions changes. Removing of the leap day set (used the calendar metadata) in the function time_slice and removed from the other functions. Vertical interpolation adjusted to the newer version of cdo and nco (tested for cdo 1.9.8 nco 4.9.4). Readign and creating parameters needed for interpolating from model to pressure levels fixed. Hardcoded parameters removed when dealing with soil variables - set to be able to read soil data from models other than CanESM2. Temporal interpolation not fixed to 6 hours any more - can be swithced to 3 hours - reads the interval from the header of the BC table.

* Vtable for CanESM2 model adjusted to be able to read the soil variabels at the lowest soil layer when running ungrib.exe.

* Reading fixed files in the metgrid section commented.

* Updated according to recent changes.

* Update run_test.sh

* Added filter, changed header

3D data on ESGF usually available on 6hr frequency. To create 3hr boundary data temoral interpolation needed.
Filter for interpolation of 6 hourly 3D data (e.g. 6hr to 3 hr) added.
Header adjusted for a specific test case, and variable "interval", denoting the frequecy of the output boundary data, added.

* Adjusted for testing historical MPIESM12LR run

* Add temporal interpolation function; add -s to several cdo commands; update vertical interpolation function

A function "Mh_to_Nh" added to interpolate Mhourly data to Nhourly data (M > N) using the cdo command "inttime" - useful if 6hourly data are available, and 3hourly output frequency is needed.
"-s" added to all cdo command to more have more clean output in the terminal/logfile while running the script (just a cosmetic change).
In the function that sets the hybrid levels "set_hybrid_levels" ap_bnds and b_bnds are kept in the output file, since cdo v1.9.8 complains if interpolation or masking is performed afterwards.

* uncomment lines

* Change filter for temporal interpolation of 6hourly variables

* Merge with Yoselita/test-patch

* Added scripts for data download

* Removed filter for fixed variables, added "remove_leap" to the header

It is not needed to interpolate in time fixed fields to run WPS without multiple runs of ungrib.exe.
"remove leap" added to the header so the preprocessor.ESGF can recognise if removing 29 Feb is desired (true) or not (false).

* Change from relative to absolute time in the output grib files

ungrib.exe is not able to read correctly realtive timestamp. Setting time units to "hours" (cdo -settunits,hours) fixes the problem.
Variable noleap is read from a BCtable. This variabels controls if the leap year will be removed or not:
For the evaluation runs typically is is not removed (false), and for the historical and scenario runs it is (true) in case the GCM uses "standard" calendar.

* Update README_BCtable.md

Added explanation of header variables and a function.

* Update README_BCtable.md

* Added "remove_leap false" to the header

* Changed and added filters, extended header

"interval" and "remove_leap" added added to the headers. Time interpolation filters renamed, and added to the surface variables

* Added function that controls leap years

Removing a leap year, if required, solved through the function "unleap". The function is included in the wrfpostfilter as the last one.

* Renamed interpolation filters

* Described new unleap filter; Renamed, removed, updated interpolation filters.

* Removed "set -u" check

This option affects variables. If variable not defined/used, error is produced, which causes the program to immediately exit.
For example, if a variable that represents land_mask (sftlf) is not processed, the script will exit producing an error "land_mask: unbound variable".
To avoid this, "set -u" has been removed.

* Added one line at the end of the BC tables

preprocessor.ESGF was exiting with failure since "while read" loop that was reading BC table was not completed correctly. Ending the BC tables with 1 line starting with "#" asures that the loop completes, and that preprocessor.ESGF exits with success if all the rest done correctly.

* Added num_metgrid_soil_levels, removed unnecessary

* Updated: Removed function format_date, comments added, reorganised

Automatic updating of the namelist.wps and namelist.input with dates corresponding to the forcing dates better formatted.
Explanation comments added.

* Changed psl interpolation filter

Interpolation filter for psl changed from fixed2interval to tinterp2interval. fixed2interval keeps only the first and the last timestep,
and interpolates the steps between. Therefore it is appropriate only for the fixed variables and those that are used only for the initilazation such as soil moisture and soil temperature.
tinterp2interval keeps all the exixting timesteps, and interpolates those that are missing.

* Updated function unleap

The script was aborting execution on error if 29 Feb was not in the file when applying unleap filter. To avoid this, dealing with 29 Feb changed.

Co-authored-by: Jesus Fernandez <jesusff@users.noreply.github.com>
Co-authored-by: zequihg50 <ezequiel.cimadevilla@unican.es>